# OpenAPI Spezifikation.yaml
...
openapi: "3.0.1"
info:
  title: "Deutscher Bundestag - DIP"
  description: "API des Dokumentations- und Informationssystems für Parlamentsmaterialien"
  version: "1.3"
  termsOfService: "https://dip.bundestag.de/%C3%BCber-dip/nutzungsbedingungen"
  contact:
    name: "Auskunfts- und Rechercheservice der Parlamentsdokumentation"
    email: "parlamentsdokumentation@bundestag.de"
servers:
- url: "https://search.dip.bundestag.de/api/v1"
security:
- ApiKeyHeader: []
- ApiKeyQuery: []
paths:
...
  /drucksache-text:
    get:
      tags:
      - "Drucksachen"
      operationId: "getDrucksacheTextList"
      summary: "Liefert eine Liste von Volltexten und Metadaten zu Drucksachen"
      parameters:
      - $ref: "#/components/parameters/AktualisiertStartFilter"
      - $ref: "#/components/parameters/AktualisiertEndFilter"
      - $ref: "#/components/parameters/DatumStartFilter"
      - $ref: "#/components/parameters/DatumEndFilter"
      - $ref: "#/components/parameters/DokumentnummerFilter"
      - $ref: "#/components/parameters/DrucksachtypFilter"
      - $ref: "#/components/parameters/IdFilter"
      - $ref: "#/components/parameters/RessortFdfFilter"
      - $ref: "#/components/parameters/TitelFilter"
      - $ref: "#/components/parameters/UrheberFilter"
      - $ref: "#/components/parameters/VorgangstypFilter"
      - $ref: "#/components/parameters/VorgangstypNotationFilter"
      - $ref: "#/components/parameters/WahlperiodeFilter"
      - $ref: "#/components/parameters/ZuordnungFilter"
      - $ref: "#/components/parameters/Cursor"
      - $ref: "#/components/parameters/Format"
      responses:
        "200":
          description: "Volltext und Metadaten zu Drucksachen"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DrucksacheTextListResponse"
            application/xml:
              schema:
                $ref: "#/components/schemas/DrucksacheTextListResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
...
  /plenarprotokoll-text:
    get:
      tags:
      - "Plenarprotokolle"
      operationId: "getPlenarprotokollTextList"
      summary: "Liefert eine Liste von Volltexten und Metadaten zu Plenarprotokollen"
      parameters:
      - $ref: "#/components/parameters/AktualisiertStartFilter"
      - $ref: "#/components/parameters/AktualisiertEndFilter"
      - $ref: "#/components/parameters/DatumStartFilter"
      - $ref: "#/components/parameters/DatumEndFilter"
      - $ref: "#/components/parameters/DokumentnummerFilter"
      - $ref: "#/components/parameters/IdFilter"
      - $ref: "#/components/parameters/VorgangstypFilter"
      - $ref: "#/components/parameters/VorgangstypNotationFilter"
      - $ref: "#/components/parameters/WahlperiodeFilter"
      - $ref: "#/components/parameters/ZuordnungFilter"
      - $ref: "#/components/parameters/Cursor"
      - $ref: "#/components/parameters/Format"
      responses:
        "200":
          description: "Volltext und Metadaten zu Plenarprotokollen"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlenarprotokollTextListResponse"
            application/xml:
              schema:
                $ref: "#/components/schemas/PlenarprotokollTextListResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
...
components:
  securitySchemes:
    ApiKeyHeader:
      type: "apiKey"
      in: "header"
      name: "Authorization"
      description: "Beispiel: *ApiKey OSOegLs.PR2lwJ1dwCeje9vTj7FPOt3hvpYKtwKkhw*\n"
    ApiKeyQuery:
      type: "apiKey"
      in: "query"
      name: "apikey"
      description: "Beispiel: *OSOegLs.PR2lwJ1dwCeje9vTj7FPOt3hvpYKtwKkhw*\n"
  parameters:
...
    DatumStartFilter:
      name: "f.datum.start"
      in: "query"
      schema:
        $ref: "#/components/schemas/Datum"
      description: "Frühestes Datum der Entität\n\nSelektiert Entitäten in einem Datumsbereich\
        \ basierend auf dem Dokumentdatum. Für Vorgänge und Personen wird der Datumsbereich\
        \ aller zugehörigen Dokumente herangezogen.\n"
      examples:
        empty:
          summary: "--"
          value: ""
        start:
          summary: "Dokumentdatum ab 11.01.2021"
          value: "2021-01-11"
    DatumEndFilter:
      name: "f.datum.end"
      in: "query"
      schema:
        $ref: "#/components/schemas/Datum"
      description: "Spätestes Datum der Entität\n\nSelektiert Entitäten in einem Datumsbereich\
        \ basierend auf dem Dokumentdatum. Für Vorgänge und Personen wird der Datumsbereich\
        \ aller zugehörigen Dokumente herangezogen.\n"
      examples:
        empty:
          summary: "--"
          value: ""
        end:
          summary: "Dokumentdatum bis einschließlich 15.01.2021"
          value: "2021-01-15"
    AktualisiertStartFilter:
      name: "f.aktualisiert.start"
      in: "query"
      schema:
        type: "string"
        format: "date-time"
      description: "Frühestes Aktualisierungsdatum der Entität\n\nSelektiert Entitä\
        ten in einem Datumsbereich basierend auf dem letzten Aktualisierungsdatum.\n"
      examples:
        empty:
          summary: "--"
          value: ""
        start:
          summary: "Aktualisierungen seit 06.12.2022 10:00 (Ortszeit Berlin)"
          value: "2022-12-06T10:00:00"
        startTimezone:
          summary: "Aktualisierungen seit 06.12.2022 10:00 (mit Zeitzone)"
          value: "2022-12-06T10:00:00+02:00"
    AktualisiertEndFilter:
      name: "f.aktualisiert.end"
      in: "query"
      schema:
        type: "string"
        format: "date-time"
      description: "Spätestes Aktualisierungsdatum der Entität\n\nSelektiert Entitä\
        ten in einem Datumsbereich basierend auf dem letzten Aktualisierungsdatum.\n"
      examples:
        empty:
          summary: "--"
          value: ""
        end:
          summary: "Aktualisierungen bis 06.12.2022 20:00 (Ortszeit Berlin)"
          value: "2022-12-06T20:00:00"
        endTimezone:
          summary: "Aktualisierungen bis 06.12.2022 20:00 (mit Zeitzone)"
          value: "2022-12-06T20:00:00+02:00"
    WahlperiodeFilter:
      name: "f.wahlperiode"
      in: "query"
      schema:
        type: "array"
        items:
          type: "integer"
      description: "Nummer der Wahlperiode\n\nSelektiert alle Entitäten, die der angegebenen\
        \ Wahlperiode zugeordnet sind.\nKann wiederholt werden, um mehrere Wahlperioden\
        \ zu selektieren. Hinterlegt ist eine ODER-Suche.\n"
...
    IdFilter:
      name: "f.id"
      in: "query"
      schema:
        type: "array"
        items:
          type: "integer"
      description: "ID der Entität\n\nKann wiederholt werden, um mehrere Entitäten\
        \ zu selektieren.\n"
...
    DokumentnummerFilter:
      name: "f.dokumentnummer"
      in: "query"
      schema:
        type: "array"
        items:
          type: "string"
          example: ""
      description: "Dokumentnummer einer Drucksache oder eines Plenarprotokolls\n\n\
        Selektiert alle Entitäten, die mit der angegebenen Dokumentnummer verknüpft\
        \ sind.\nKann wiederholt werden, um mehrere Dokumentnummern zu selektieren.\
        \ Hinterlegt ist eine ODER-Suche.\n"
...
    DrucksachtypFilter:
      name: "f.drucksachetyp"
      in: "query"
      schema:
        type: "string"
      description: "Typ der Drucksache\n\nSelektiert alle Entitäten, die mit dem angegebenen\
        \ Drucksachetyp verknüpft sind.\n"
      examples:
        none:
          summary: "--"
          value: ""
        single:
          summary: "Antrag"
          value: "Antrag"
...
    VorgangstypFilter:
      name: "f.vorgangstyp"
      in: "query"
      schema:
        type: "array"
        items:
          type: "string"
          example: ""
      description: "Vorgangstyp\n\nSelektiert alle Entitäten, die dem angegebenen\
        \ Vorgangstyp zugeordnet sind.\nKann wiederholt werden, um mehrere Vorgangstypen\
        \ zu selektieren. Hinterlegt ist eine ODER-Suche.\n"
      examples:
        none:
          summary: "--"
          value: ""
        single:
          summary: "Gesetzgebung"
          value:
          - "Gesetzgebung"
    VorgangstypNotationFilter:
      name: "f.vorgangstyp_notation"
      in: "query"
      schema:
        type: "array"
        items:
          type: "integer"
      description: "Vorgangstyp-Notation\n\nSelektiert alle Entitäten, die der angegebenen\
        \ Vorgangstyp-Notation zugeordnet sind.\nKann wiederholt werden, um mehrere\
        \ Vorgangstypen zu selektieren. Hinterlegt ist eine ODER-Suche.\n"
      examples:
        none:
          summary: "--"
          value: ""
        single:
          summary: "100 (Gesetzgebung)"
          value:
          - 100
...
    RessortFdfFilter:
      name: "f.ressort_fdf"
      in: "query"
      schema:
        type: "array"
        items:
          type: "string"
          example: ""
      description: "Ressort (federführend)\n\nSelektiert alle Entitäten, die mit dem\
        \ angegebenen federführenden Ressort in einer Drucksache verknüpft sind.\n\
        Kann wiederholt werden, um die Schnittmenge mehrerer Ressorts zu selektieren.\
        \ Hinterlegt ist eine UND-Suche. \nEine ODER-Suche über mehrere federführende\
        \ Ressorts ist mit einer einzigen Abfrage nicht möglich. Die federführenden\
        \ Ressorts müssen dazu einzeln abgefragt werden.\n"
      examples:
        none:
          summary: "--"
          value: ""
        single:
          summary: "Bundesministerium für Gesundheit"
          value:
          - "Bundesministerium für Gesundheit"
        multiple:
          summary: "Bundesministerium für Gesundheit UND Bundesministerium für Arbeit\
            \ und Soziales"
          value:
          - "Bundesministerium für Gesundheit"
          - "Bundesministerium für Arbeit und Soziales"
...
    TitelFilter:
      name: "f.titel"
      in: "query"
      schema:
        type: "array"
        items:
          type: "string"
          example: ""
      description: "Titel\n\nSelektiert alle Entitäten, die den angegebenen Suchbegriff\
        \ im Titel enthalten.\nKann wiederholt werden, um mehrere Titel zu selektieren.\
        \ Hinterlegt ist eine ODER-Suche.\nEine Einzelwortsuche ist möglich. Mehrere\
        \ Suchbegriffe hintereinander werden als Phrase gesucht.\n"
      examples:
        none:
          summary: "--"
          value: ""
        single:
          summary: "Veröffentlichung von Spenden, die im Einzelfall die Höhe von 50\
            \ 000 Euro übersteigen"
          value:
          - "Veröffentlichung von Spenden, die im Einzelfall die Höhe von 50 000 Euro\
            \ übersteigen"
        multiple:
          summary: "Pflegehilfe ODER Pflegesatz"
          value:
          - "Pflegehilfe"
          - "Pflegesatz"
    UrheberFilter:
      name: "f.urheber"
      in: "query"
      schema:
        type: "array"
        items:
          type: "string"
          example: ""
      description: "Urheber\n\nSelektiert alle Entitäten, die mit dem angegebenen\
        \ Urheber in einer Drucksache verknüpft sind.\nKann wiederholt werden, um\
        \ die Schnittmenge mehrerer Urheber zu selektieren. Hinterlegt ist eine UND-Suche.\
        \ \nEine ODER-Suche über mehrere Urheber ist mit einer einzigen Abfrage nicht\
        \ möglich. Die Urheber müssen dazu einzeln abgefragt werden.\n"
      examples:
        none:
          summary: "--"
          value: ""
        single:
          summary: "Bundesregierung"
          value:
          - "Bundesregierung"
        multiple:
          summary: "Fraktion der SPD UND Fraktion Bündnis 90/Die Grünen"
          value:
          - "Fraktion der SPD"
          - "Fraktion Bündnis 90/Die Grünen"
    ZuordnungFilter:
      name: "f.zuordnung"
      in: "query"
      schema:
        $ref: "#/components/schemas/Zuordnung"
      description: "Zuordnung der Entität zum Bundestag, Bundesrat, Bundesversammlung\
        \ oder Europakammer"
    Format:
      name: "format"
      in: "query"
      schema:
        type: "string"
        enum:
        - "json"
        - "xml"
        default: "json"
      description: "Steuert das Datenformat der Antwort, möglich sind JSON (voreingestellt)\
        \ oder XML."
    Cursor:
      name: "cursor"
      in: "query"
      schema:
        type: "string"
      description: "Position des Cursors zur Anfrage weiterer Entitäten\n\nÜbersteigt\
        \ die Anzahl der gefundenen Entitäten das jeweilige Limit, muss eine Folgeanfrage\
        \ gestellt werden, um weitere Entitäten zu laden.\nEine Folgeanfrage wird\
        \ gebildet, indem alle Parameter der ursprünglichen Anfrage wiederholt werden\
        \ und zusätzlich der cursor Parameter der letzten Antwort eingesetzt wird.\n\
        Es können solange Folgeanfragen gestellt werden, bis sich der cursor nicht\
        \ mehr ändert. Dies signalisiert, dass alle Entitäten geladen wurden.\n"
  responses:
    BadRequestResponse:
      description: "Syntaxfehler in einem der Anfrageparameter"
      content:
        application/json:
          schema:
            type: "object"
            properties:
              code:
                type: "integer"
                enum:
                - 400
              message:
                type: "string"
                example: "Invalid cursor"
            required:
            - "code"
            - "message"
    UnauthorizedResponse:
      description: "Ein gültiger API-Key ist für alle Anfragen erforderlich. Dieser\
        \ kann entweder im HTTP Authorization Header oder als Anfrageparameter apikey\
        \ gesendet werden."
      content:
        application/json:
          schema:
            type: "object"
            properties:
              code:
                type: "integer"
                enum:
                - 401
              message:
                type: "string"
                enum:
                - "An API key is required to access this service. Please refer to\
                  \ https://dip.bundestag.de/über-dip/hilfe/api how to apply for a\
                  \ key. Misuse of this service may lead to blocking your requests."
            required:
            - "code"
            - "message"
...
  schemas:
    ListResponseBase:
      type: "object"
      xml:
        name: "response"
      properties:
        numFound:
          type: "integer"
          format: "int32"
          example: 176
        cursor:
          type: "string"
          example: "AoJwgNjC_PYCMURydWNrc2FjaGUtMjQ5MjYw"
      required:
      - "numFound"
      - "cursor"
...
    PlenarprotokollTextListResponse:
      allOf:
      - $ref: "#/components/schemas/ListResponseBase"
      - type: "object"
        properties:
          documents:
            type: "array"
            items:
              $ref: "#/components/schemas/PlenarprotokollText"
            minItems: 0
            maxItems: 10
        required:
        - "documents"
    DrucksacheTextListResponse:
      allOf:
      - $ref: "#/components/schemas/ListResponseBase"
      - type: "object"
        properties:
          documents:
            type: "array"
            items:
              $ref: "#/components/schemas/DrucksacheText"
            minItems: 0
            maxItems: 10
        required:
        - "documents"
...
    Drucksache:
      description: "Liefert Metadaten zu einer Drucksache."
      type: "object"
      xml:
        name: "document"
      properties:
        id:
          type: "string"
          pattern: "^\\d+$"
          example: 68852
        typ:
          type: "string"
          enum:
          - "Dokument"
          example: "Dokument"
        dokumentart:
          type: "string"
          enum:
          - "Drucksache"
          example: "Drucksache"
        drucksachetyp:
          type: "string"
          example: "Antrag"
        dokumentnummer:
          type: "string"
          example: "19/1"
        wahlperiode:
          type: "integer"
          format: "int32"
          example: 19
        herausgeber:
          type: "string"
          enum:
          - "BT"
          - "BR"
          example: "BT"
        datum:
          type: "string"
          format: "date"
          example: "2017-10-24"
        aktualisiert:
          type: "string"
          format: "date-time"
          example: "2022-08-01T15:30:16+02:00"
          description: "Letzte Aktualisierung der Entität"
        titel:
          type: "string"
          example: "Weitergeltung von Geschäftsordnungsrecht"
        autoren_anzeige:
          type: "array"
          items:
            type: "object"
            properties:
              id:
                type: "string"
                pattern: "^\\d+$"
                example: 546
                description: "ID von Personenstammdaten"
              title:
                type: "string"
                example: "Kai Gehring, MdB, BÜNDNIS 90/DIE GRÜNEN"
              autor_titel:
                type: "string"
                example: "Kai Gehring"
            required:
            - "id"
            - "title"
            - "autor_titel"
          minItems: 0
          maxItems: 4
          description: "Zusammenfassung der ersten 4 zur Anzeige markierten Autor:innen"
        autoren_anzahl:
          type: "integer"
          format: "int32"
          example: 4
          description: "Gesamtzahl der Autor:innen"
        fundstelle:
          $ref: "#/components/schemas/Fundstelle"
        pdf_hash:
          type: "string"
          example: "a33af31e7c4524db8db172ef8f9e0f6d"
          description: "MD5-Prüfsumme der PDF-Datei"
        urheber:
          type: "array"
          items:
            $ref: "#/components/schemas/Urheber"
        vorgangsbezug:
          type: "array"
          items:
            $ref: "#/components/schemas/Vorgangsbezug"
          minItems: 0
          maxItems: 4
          description: "Zusammenfassung der ersten 4 zugehörigen Vorgänge"
        vorgangsbezug_anzahl:
          type: "integer"
          format: "int32"
          example: 6
          description: "Gesamtzahl der zugehörigen Vorgänge"
        ressort:
          type: "array"
          items:
            $ref: "#/components/schemas/Ressort"
        anlagen:
          type: "string"
          example: "Verzeichnis der Gutachten des Sachverständigenrates"
      required:
      - "id"
      - "drucksachetyp"
      - "dokumentart"
      - "autoren_anzahl"
      - "typ"
      - "vorgangsbezug_anzahl"
      - "dokumentnummer"
      - "herausgeber"
      - "datum"
      - "aktualisiert"
      - "titel"
      - "fundstelle"
    Plenarprotokoll:
      description: "Liefert Metadaten zu einem Plenarprotokoll."
      type: "object"
      xml:
        name: "document"
      properties:
        id:
          type: "string"
          pattern: "^\\d+$"
          example: 908
        dokumentart:
          type: "string"
          enum:
          - "Plenarprotokoll"
          example: "Plenarprotokoll"
        typ:
          type: "string"
          enum:
          - "Dokument"
          example: "Dokument"
        dokumentnummer:
          type: "string"
          example: "19/1"
        wahlperiode:
          type: "integer"
          format: "int32"
          example: 19
        herausgeber:
          $ref: "#/components/schemas/Zuordnung"
        datum:
          type: "string"
          format: "date"
          example: "2017-10-24"
        aktualisiert:
          type: "string"
          format: "date-time"
          example: "2022-08-01T15:30:16+02:00"
          description: "Letzte Aktualisierung der Entität"
        titel:
          type: "string"
          example: "Protokoll der 1. Sitzung des 19. Deutschen Bundestages"
        fundstelle:
          $ref: "#/components/schemas/Fundstelle"
        pdf_hash:
          type: "string"
          example: "a33af31e7c4524db8db172ef8f9e0f6d"
          description: "MD5-Prüfsumme der PDF-Datei"
        vorgangsbezug:
          type: "array"
          items:
            $ref: "#/components/schemas/Vorgangsbezug"
          minItems: 0
          maxItems: 4
          description: "Zusammenfassung der ersten 4 zugehörigen Vorgänge"
        vorgangsbezug_anzahl:
          type: "integer"
          format: "int32"
          example: 6
          description: "Gesamtzahl der zugehörigen Vorgänge"
        sitzungsbemerkung:
          type: "string"
          example: "Sondersitzung"
      required:
      - "id"
      - "dokumentart"
      - "typ"
      - "vorgangsbezug_anzahl"
      - "dokumentnummer"
      - "herausgeber"
      - "datum"
      - "aktualisiert"
      - "titel"
      - "fundstelle"
    DokumentTextBase:
      description: "Liefert zusätzlich zu den Metadaten den Volltext eines Dokuments."
      type: "object"
      properties:
        text:
          type: "string"
          description: "Volltext des Dokuments\n\nDas Beispiel enthält einen gekü\
            rzten Auszug einer Drucksache.\n"
          example: "Deutscher Bundestag Drucksache 19/1\n19. Wahlperiode 24.10.2017\n\
            \nAntrag\nder Fraktion der CDU/CSU\nWeitergeltung von Geschäftsordnungsrecht\n\
            Der Bundestag wolle beschließen:\nFür die 19. Wahlperiode werden übernommen\n\
            – die Geschäftsordnung des Deutschen Bundestages einschließlich ihrer\
            \ Anla-\ngen, soweit sie vom Deutschen Bundestag zu beschließen sind,\
            \ in der Fassung\nder Bekanntmachung vom 2. Juli 1980 (BGBl. I S. 1237),\
            \ zuletzt geändert\nlaut Bekanntmachung vom 12. Juni 2017 (BGBl. I S.\
            \ 1877), wobei § 126a der\nGeschäftsordnung mit Ablauf der 18. Wahlperiode\
            \ entfallen ist;\n– die Gemeinsame Geschäftsordnung des Bundestages und\
            \ des Bundesrates für\nden Ausschuss nach Artikel 77 des Grundgesetzes\
            \ (Vermittlungsausschuss)\nvom 5. Mai 1951 (BGBl. II S. 103), zuletzt\
            \ geändert laut Bekanntmachung vom\n30. April 2003 (BGBl. I S. 677);\n\
            – die Geschäftsordnung für den Gemeinsamen Ausschuss vom 23. Juli 1969\n\
            (BGBl. I S. 1102), zuletzt geändert laut Bekanntmachung vom 20. Juli 1993\n\
            (BGBl. I S. 1500);\n– die Geschäftsordnung für das Verfahren nach Artikel\
            \ 115d des Grundgesetzes\nvom 23. Juli 1969 (BGBl. I S. 1100);\n– die\
            \ Richtlinien zur Überprüfung auf eine Tätigkeit oder politische Verantwor-\n\
            tung für das Ministerium für Staatssicherheit/Amt für Nationale Sicherheit\
            \ der\nehemaligen Deutschen Demokratischen Republik vom 13. Dezember 1991\n\
            (BGBl. 1992 I S. 76), zuletzt geändert laut Bekanntmachung vom 21. Oktober\n\
            2005 (BGBl. I S. 3094).\nBerlin, den 23. Oktober 2017\nVolker Kauder,\
            \ Alexander Dobrindt und Fraktion\nSatz: Satzweiss.com Print, Web, Software\
            \ GmbH, Mainzer Straße 116, 66121 Saarbrücken, www.satzweiss.com\nDruck:\
            \ Printsystem GmbH, Schafwäsche 1-3, 71296 Heimsheim, www.printsystem.de\n\
            Vertrieb: Bundesanzeiger Verlag GmbH, Postfach 10 05 34, 50445 Köln, Telefon\
            \ (02 21) 97 66 83 40, Fax (02 21) 97 66 83 44, www.betrifft-gesetze.de\n\
            ISSN 0722-8333\n"
    DrucksacheText:
      allOf:
      - $ref: "#/components/schemas/Drucksache"
      - $ref: "#/components/schemas/DokumentTextBase"
    PlenarprotokollText:
      allOf:
      - $ref: "#/components/schemas/Plenarprotokoll"
      - $ref: "#/components/schemas/DokumentTextBase"
...
    Fundstelle:
      description: "Liefert im Vorgangsablauf das zu einem Vorgangsschritt gehörende\
        \ Dokument (Drucksache oder Protokoll). \n\nBeispiel: „BT-Drucksache 19/1\
        \ (Antrag Fraktion der CDU/CSU)“ oder beim Vorgangsschritt Beratung „BT-Plenarprotokoll\
        \ 19/1, S. 4C-12A“.\n"
      type: "object"
      properties:
        id:
          type: "string"
          pattern: "^\\d+$"
          example: 68852
          description: "ID einer Drucksache oder eines Plenarprotokolls"
        dokumentart:
          type: "string"
          enum:
          - "Drucksache"
          - "Plenarprotokoll"
          example: "Drucksache"
        pdf_url:
          type: "string"
          example: "https://dserver.bundestag.de/btd/19/000/1900001.pdf"
        dokumentnummer:
          type: "string"
          example: "19/1"
        datum:
          type: "string"
          format: "date"
          example: "2017-10-24"
        drucksachetyp:
          type: "string"
          example: "Antrag"
        herausgeber:
          $ref: "#/components/schemas/Zuordnung"
        urheber:
          type: "array"
          items:
            type: "string"
            example: "Fraktion der CDU/CSU"
            description: "Körperschaftliche Urheber der zugehörigen Drucksache"
        verteildatum:
          type: "string"
          format: "date"
          example: "2017-10-25"
        seite:
          type: "string"
          example: "4292B"
        anfangsseite:
          type: "integer"
          example: 4251
        endseite:
          type: "integer"
          example: 4298
        anfangsquadrant:
          $ref: "#/components/schemas/Quadrant"
        endquadrant:
          $ref: "#/components/schemas/Quadrant"
        frage_nummer:
          type: "string"
          example: "C.4"
        anlagen:
          type: "string"
          example: "Verzeichnis der Gutachten des Sachverständigenrates"
        top:
          type: "integer"
          format: "int32"
          example: 2
        top_zusatz:
          type: "string"
          example: "b"
      required:
      - "id"
      - "dokumentnummer"
      - "datum"
      - "dokumentart"
      - "herausgeber"
      - "urheber"
...
    Ressort:
      description: "Liefert das Ressort (Bundesministerium), das innerhalb der Bundesregierung\
        \ federführend für die Ausarbeitung z. B. eines Gesetzentwurfs oder eines\
        \ Antrags oder die Beantwortung von Anfragen ist. Wird nur verwendet, wenn\
        \ Urheber einer Drucksache die Bundesregierung ist.\n\nDas Ressort wird auch\
        \ bei persönlichen Urhebern verwendet, um die Funktion von Angehörigen der\
        \ Bundesregierung zu präzisieren, z. B. „Dr. Robert Habeck, Bundesmin., Bundesministerium\
        \ für Wirtschaft und Klimaschutz“ oder „Dr. Florian Toncar, Parl. Staatssekr.,\
        \ Bundesministerium der Finanzen“.\n"
      type: "object"
      properties:
        federfuehrend:
          type: "boolean"
          example: true
        titel:
          type: "string"
          example: "Bundesministerium für Wirtschaft und Energie"
      required:
      - "federfuehrend"
      - "titel"
...
    Urheber:
      description: "Liefert den körperschaftlichen Urheber einer Bundestags- oder\
        \ Bundesratsdrucksache, z. B. einen Ausschuss, eine Fraktion, die Bundesregierung,\
        \ ein Bundesland oder dergleichen."
      type: "object"
      properties:
        einbringer:
          type: "boolean"
        bezeichnung:
          type: "string"
          example: "B90/GR"
        titel:
          type: "string"
          example: "Fraktion BÜNDNIS 90/DIE GRÜNEN"
        rolle:
          type: "string"
          enum:
          - "B"
          - "U"
      required:
      - "bezeichnung"
      - "titel"
...
    Vorgangsbezug:
      description: "Liefert ID, Titel und Vorgangstyp eines Vorgangs, der mit der\
        \ Drucksache oder dem Plenarprotokoll verbunden ist."
      type: "object"
      properties:
        id:
          type: "string"
          pattern: "^\\d+$"
          example: 84393
          description: "ID eines verknüpften Vorgangs"
        titel:
          type: "string"
          example: "Eröffnung der 1. Sitzung des 19. Deutschen Bundestages"
        vorgangstyp:
          type: "string"
          example: "Ansprache/Erklärung/Mitteilung"
      required:
      - "id"
      - "titel"
      - "vorgangstyp"
...
    Datum:
      description: "Liefert das Datum eines Dokuments (Drucksache oder Protokoll)."
      type: "string"
      format: "date"
    Quadrant:
      description: "Teil der Fundstelle eines Plenarprotokolls. Jede Seite im Plenarprotokoll\
        \ ist in vier gleich große Viertel unterteilt (Quadranten) mit den Bezeichnungen\
        \ A, B, C, D."
      type: "string"
      enum:
      - "A"
      - "B"
      - "C"
      - "D"
    Zuordnung:
      description: "Jeder Vorgangsschritt ist entweder dem Bundestag (BT), dem Bundesrat\
        \ (BR), der Bundesversammlung (BV) oder der Europakammer (EK) zugeordnet.\
        \ Über die Zuordnung lassen sich bspw. Rechtsverordnungen herausfiltern, an\
        \ denen der Bundestag beteiligt / nicht beteiligt war."
      type: "string"
      enum:
      - "BT"
      - "BR"
      - "BV"
      - "EK"